Fix pygbag compatibility using asyncio and web-safe main loop
